Move stats querying into own module
